refactor: split `dns` module
